Add Range::front() + back(). Fix end iterator
Previously, the only way to get the Range front and back was through the
iterators. I've now added front() and back() functions to do
this properly.

Additionally, the end() iterator wasn't one past the element, meaning it
never included the last element.

For more advanced ranges, the std::ranges library is recommended.